Create USD Exporter
